requirements.txt
alabaster==0.7.12
aniso8601==8.0.0
asn1crypto==1.3.0
attrs==19.3.0
autoapi==2.0.1
Babel==2.8.0
backcall==0.1.0
bleach==3.1.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
get==2019.4.13
grip==4.5.2
idna==2.9
imagesize==1.2.0
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
path-and-address==2.0.1
pexpect==4.8.0
pickleshare==0.7.5
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
public==2019.4.13
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
query-string==2019.4.13
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==1.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==3.1.0
